netsuite-ai-webapp/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.27.1
pydantic-settings==2.7.1
python-json-logger==3.2.1
sqlalchemy==2.0.30
alembic==1.13.2
psycopg[binary]==3.2.4
redis==5.2.0
cryptography==43.0.3
jaydebeapi==1.2.3
JPype1==1.5.2
openai==1.59.3
openpyxl==3.1.5
reportlab==4.2.5
